Bugfix due to duplicate classes
It appears that Godot may be automatically generating classes somewhere from filenames, as such declaring them identically within the scripts themselves causes conflicts. Either this or it's a weird quick of the system, but either way, this seems to have cleared it up on my end at least.
Someone may want to open my branch and run it on their machine just to make sure all is well before merging.